move swagger definiiton to separate file
